app/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
redis==5.0.3
python-dotenv==1.0.1
Pillow